Updated to match traditional server file structure
Azure will only recognize the node.js server if it is at root (I don't
know how to specify subdirectories). Added a root endpoint for the
server to return the client html code.

Renamed "/Client" to "/public" to be consistent with standard
conventions).

Updated server urls main.js to reflect the azure url.